Clean up BMCWEB_ENABLE_SSL
This macro came originally from CROW_ENABLE_SSL, and was used as a macro
to optionally compile without openssl being required.

OpenSSL has been pulled into many other dependencies, and has been
functionally required to be included for a long time, so there's no
reason to hold onto this macro.

Remove most uses of the macro, and for the couple functional places the
macro is used, transition to a constexpr if to enable the TLS paths.

Tested: Redfish service validator passes.

+++ http/http_server.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "bmcweb_config.h"
+
 #include "http_connection.hpp"
 #include "logging.hpp"
 #include "ssl_key_handler.hpp"
@@ -27,8 +29,7 @@ template <typename Handler, typename Adaptor = boost::asio::ip::tcp::socket>
 class Server
 {
   public:
-    Server(Handler* handlerIn,
-           std::unique_ptr<boost::asio::ip::tcp::acceptor>&& acceptorIn,
+    Server(Handler* handlerIn, boost::asio::ip::tcp::acceptor&& acceptorIn,
            std::shared_ptr<boost::asio::ssl::context> adaptorCtxIn,
            std::shared_ptr<boost::asio::io_context> io =
                std::make_shared<boost::asio::io_context>()) :
@@ -36,27 +37,6 @@ class Server
         acceptor(std::move(acceptorIn)),
         signals(*ioService, SIGINT, SIGTERM, SIGHUP), handler(handlerIn),
         adaptorCtx(std::move(adaptorCtxIn))
-    {}
-
-    Server(Handler* handlerIn, uint16_t port,
-           const std::shared_ptr<boost::asio::ssl::context>& adaptorCtxIn,
-           const std::shared_ptr<boost::asio::io_context>& io =
-               std::make_shared<boost::asio::io_context>()) :
-        Server(handlerIn,
-               std::make_unique<boost::asio::ip::tcp::acceptor>(
-                   *io, boost::asio::ip::tcp::endpoint(
-                            boost::asio::ip::make_address("0.0.0.0"), port)),
-               adaptorCtxIn, io)
-    {}
-
-    Server(Handler* handlerIn, int existingSocket,
-           const std::shared_ptr<boost::asio::ssl::context>& adaptorCtxIn,
-           const std::shared_ptr<boost::asio::io_context>& io =
-               std::make_shared<boost::asio::io_context>()) :
-        Server(handlerIn,
-               std::make_unique<boost::asio::ip::tcp::acceptor>(
-                   *io, boost::asio::ip::tcp::v6(), existingSocket),
-               adaptorCtxIn, io)
     {}
 
     void updateDateStr()
@@ -90,47 +70,48 @@ class Server
         };
 
         BMCWEB_LOG_INFO("bmcweb server is running, local endpoint {}",
-                        acceptor->local_endpoint().address().to_string());
+                        acceptor.local_endpoint().address().to_string());
         startAsyncWaitForSignal();
         doAccept();
     }
 
     void loadCertificate()
     {
-#ifdef BMCWEB_ENABLE_SSL
-        namespace fs = std::filesystem;
-        // Cleanup older certificate file existing in the system
-        fs::path oldCert = "/home/root/server.pem";
-        if (fs::exists(oldCert))
+        if constexpr (bmcwebEnableTLS)
         {
-            fs::remove("/home/root/server.pem");
+            namespace fs = std::filesystem;
+            // Cleanup older certificate file existing in the system
+            fs::path oldCert = "/home/root/server.pem";
+            if (fs::exists(oldCert))
+            {
+                fs::remove("/home/root/server.pem");
+            }
+            fs::path certPath = "/etc/ssl/certs/https/";
+            // if path does not exist create the path so that
+            // self signed certificate can be created in the
+            // path
+            if (!fs::exists(certPath))
+            {
+                fs::create_directories(certPath);
+            }
+            fs::path certFile = certPath / "server.pem";
+            BMCWEB_LOG_INFO("Building SSL Context file={}", certFile.string());
+            std::string sslPemFile(certFile);
+            std::string cert =
+                ensuressl::ensureOpensslKeyPresentAndValid(sslPemFile);
+            if (cert.empty())
+            {
+                throw std::runtime_error("Failed to load string");
+            }
+            std::shared_ptr<boost::asio::ssl::context> sslContext =
+                ensuressl::getSslContext(cert);
+            if (sslContext == nullptr)
+            {
+                throw std::runtime_error("Failed to load certificate");
+            }
+            adaptorCtx = sslContext;
+            handler->ssl(std::move(sslContext));
         }
-        fs::path certPath = "/etc/ssl/certs/https/";
-        // if path does not exist create the path so that
-        // self signed certificate can be created in the
-        // path
-        if (!fs::exists(certPath))
-        {
-            fs::create_directories(certPath);
-        }
-        fs::path certFile = certPath / "server.pem";
-        BMCWEB_LOG_INFO("Building SSL Context file={}", certFile.string());
-        std::string sslPemFile(certFile);
-        std::string cert =
-            ensuressl::ensureOpensslKeyPresentAndValid(sslPemFile);
-        if (cert.empty())
-        {
-            throw std::runtime_error("Failed to load string");
-        }
-        std::shared_ptr<boost::asio::ssl::context> sslContext =
-            ensuressl::getSslContext(cert);
-        if (sslContext == nullptr)
-        {
-            throw std::runtime_error("Failed to load certificate");
-        }
-        adaptorCtx = sslContext;
-        handler->ssl(std::move(sslContext));
-#endif
     }
 
     void startAsyncWaitForSignal()
@@ -148,7 +129,7 @@ class Server
                     BMCWEB_LOG_INFO("Receivied reload signal");
                     loadCertificate();
                     boost::system::error_code ec2;
-                    acceptor->cancel(ec2);
+                    acceptor.cancel(ec2);
                     if (ec2)
                     {
                         BMCWEB_LOG_ERROR(
@@ -172,12 +153,24 @@ class Server
 
     void doAccept()
     {
+        if (ioService == nullptr)
+        {
+            BMCWEB_LOG_CRITICAL("ioService was null?");
+            return;
+        }
+        BMCWEB_LOG_DEBUG("Starting accept");
         boost::asio::steady_timer timer(*ioService);
         std::shared_ptr<Connection<Adaptor, Handler>> connection;
         if constexpr (std::is_same<Adaptor,
                                    boost::beast::ssl_stream<
                                        boost::asio::ip::tcp::socket>>::value)
         {
+            if (adaptorCtx == nullptr)
+            {
+                BMCWEB_LOG_CRITICAL("TLS context was null????");
+                return;
+            }
+
             connection = std::make_shared<Connection<Adaptor, Handler>>(
                 handler, std::move(timer), getCachedDateStr,
                 Adaptor(*ioService, *adaptorCtx));
@@ -188,11 +181,17 @@ class Server
                 handler, std::move(timer), getCachedDateStr,
                 Adaptor(*ioService));
         }
-        acceptor->async_accept(
+        acceptor.async_accept(
             boost::beast::get_lowest_layer(connection->socket()),
             [this, connection](const boost::system::error_code& ec) {
-            if (!ec)
+            if (ec)
             {
+                BMCWEB_LOG_ERROR("Async accept hit error");
+            }
+            else
+            {
+                BMCWEB_LOG_DEBUG("{} Starting socket",
+                                 logPtr(connection.get()));
                 boost::asio::post(*ioService,
                                   [connection] { connection->start(); });
             }
@@ -203,7 +202,7 @@ class Server
   private:
     std::shared_ptr<boost::asio::io_context> ioService;
     std::function<std::string()> getCachedDateStr;
-    std::unique_ptr<boost::asio::ip::tcp::acceptor> acceptor;
+    boost::asio::ip::tcp::acceptor acceptor;
     boost::asio::signal_set signals;
 
     std::string dateStr;
